result module is renamed as fallible
