+++ src/computational_variants/fallible/map_fallible.rs
