+++ src/computational_variants/map.rs
@@ -1,6 +1,6 @@
 use super::xap::ParXap;
 use crate::ParIterFallible;
-use crate::computational_variants::result::ParMapFallible;
+use crate::computational_variants::fallible::ParMapFallible;
 use crate::par_iter_fallible::IntoResult;
 use crate::values::{Vector, WhilstAtom};
 use crate::{
